refactor: move cd command handling to a separate module
