+++ library/changeDirectory.js
@@ -1,6 +1,7 @@
 import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 
+
 export async function changeDirectory(currentDir, targetDir) {
   if (!targetDir) {
     console.log('Operation failed. You need to specify the directory');
